requirements.txt
pandas
scikit-learn
joblib
streamlit
fastapi
python-dotenv
google-generativeai
